Introduce the service definition label to the existing tests
This change is amending to the existing tests describe nodes.
Which allows the suite of tests to be executed using ginkgos label
filter feature over using the focus filter.

Focus filter will continue to work as the string Suite:
service-definition still exists within the test case names.
It will eventually be removed in replacement of the label identifiers.