Add support to use aws profile for rosa and aws authentication
This commit adds support to osde2e users allowing them to use aws
profiles to authenticate with aws. This gies an additional option over
just using access keys for authentication.

When the profile is set, access keys can be ignored, when profile is not set
access keys will need to be set to authenticate.